requirements.txt
asgiref==3.3.1
Django==3.1.6
django-ckeditor==6.0.0
django-ckeditor-updated==4.4.4
django-cleanup==5.1.0
django-jalali-date==0.3.1
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-render-partial==0.4
jdatetime==3.6.2
phonenumbers==8.12.18
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
